requirements.txt
streamlit
pandas
numpy
google-generativeai
scikit-learn
sentence-transformers
